+++ R/rhat.R
@@ -1,7 +1,8 @@
 #' R-hat
 #'
-#' By default calculates the uncorrected, univariate, untransformed,
-#' split R-hat (potential scale reduction factor) values.
+#' By default calculates the maximum of the uncorrected,
+#' split, folded and unfolded univariate,
+#' rank normalized R-hat (potential scale reduction factor) values.
 #'
 #' @param x An MCMC object.
 #' @param ... Unused.
@@ -9,8 +10,7 @@
 #' "term", "parameter" or "all".
 #' @param as_df A flag indicating whether to return the values as a
 #' data frame versus a named list.
-#' @param split A flag specifying whether to split the chains.
-#' @param folded A flag specifying whether to convert values into their absolute deviations from the global median.
+#' @param folded A logical scalar specifying whether to convert values into their absolute deviations from the global median or if NA do both and select the maximum.
 #' @param normalized A flag specifying whether to rank normalize the values.
 #' @param bound flag specifying whether to bind mcmcrs objects by their chains before calculating rhat.
 #' @return The rhat value(s).
@@ -30,42 +30,39 @@ rhat <- function(x, ...) {
 
 #' @describeIn rhat R-hat for an mcarray object
 #' @export
-rhat.mcarray <- function(x, by = "all", as_df = FALSE,
-                         split = TRUE, folded = FALSE, normalized = FALSE,
+rhat.mcarray <- function(x, by = "all", as_df = FALSE, folded = NA, normalized = TRUE,
                          ...) {
   check_unused(...)
-  rhat(as.mcmcarray(x), by = by, as_df = as_df,
-       split = split, folded = folded, normalized = normalized)
+  rhat(as.mcmcarray(x), by = by, as_df = as_df, folded = folded, normalized = normalized)
 }
 
 #' @describeIn rhat R-hat for an mcmc object
 #' @export
-rhat.mcmc <- function(x, by = "all", as_df = FALSE,
-                      split = TRUE, folded = FALSE, normalized = FALSE, ...) {
+rhat.mcmc <- function(x, by = "all", as_df = FALSE, folded = NA, normalized = TRUE, ...) {
   check_unused(...)
   rhat(as.mcmcr(x), by = by, as_df = as_df,
-       split = split, folded = folded, normalized = normalized)
+       folded = folded, normalized = normalized)
 }
 
 #' @describeIn rhat R-hat for an mcmc.list object
 #' @export
 rhat.mcmc.list <- function(x, by = "all", as_df = FALSE,
-                           split = TRUE, folded = FALSE, normalized = FALSE,
+                           folded = NA, normalized = TRUE,
                            ...) {
   check_unused(...)
   rhat(as.mcmcr(x), by = by, as_df = as_df,
-       split = split, folded = folded, normalized = normalized)
+       folded = folded, normalized = normalized)
 }
 
 #' @describeIn rhat R-hat for an mcmcarray object
 #' @export
 rhat.mcmcarray <- function(x, by = "all", as_df = FALSE,
-                           split = TRUE, folded = FALSE, normalized = FALSE,
+                           folded = NA, normalized = TRUE,
                            ...) {
   check_unused(...)
   check_vector(by, c("all", "parameter", "term"), length = 1)
   check_flag(as_df)
-  check_flag(split)
+  check_scalar(folded, c(TRUE, NA))
   check_flag(normalized)
 
   if(niters(x) < 4) {
@@ -78,8 +75,13 @@ rhat.mcmcarray <- function(x, by = "all", as_df = FALSE,
     return(tibble(parameter = "parameter", rhat = NA_real_))
   }
 
-  if(split) x <- split_chains(x)
-  x <- apply(x, 3:ndims(x), FUN = .rhat, folded = folded, normalized = normalized)
+  x <- split_chains(x)
+  if(is.na(folded)) {
+    x_folded <- apply(x, 3:ndims(x), FUN = .rhat, folded = TRUE, normalized = normalized)
+    x_unfolded <- apply(x, 3:ndims(x), FUN = .rhat, folded = FALSE, normalized = normalized)
+    x <- pmax(x_folded, x_unfolded)
+  } else
+    x <- apply(x, 3:ndims(x), FUN = .rhat, folded = folded, normalized = normalized)
 
   if(!as_df) {
     if(by == "term") return(x)
@@ -95,11 +97,11 @@ rhat.mcmcarray <- function(x, by = "all", as_df = FALSE,
 #' @describeIn rhat R-hat for an mcmcr object
 #' @export
 rhat.mcmcr <- function(x, by = "all", as_df = FALSE,
-                       split = TRUE, folded = FALSE, normalized = FALSE, ...) {
+                       folded = NA, normalized = TRUE, ...) {
   check_unused(...)
   parameters <- parameters(x)
   x <- lapply(x, rhat, by = by, as_df = as_df,
-              split = split, folded = folded, normalized = normalized)
+              folded = folded, normalized = normalized)
   if(!as_df) {
     if (by != "all") return(x)
     return(max(unlist(x)))
@@ -114,8 +116,8 @@ rhat.mcmcr <- function(x, by = "all", as_df = FALSE,
 #' @describeIn rhat R-hat for an mcmcrs object
 #' @export
 rhat.mcmcrs <- function(x, by = "all", as_df = FALSE,
-                        split = !bound, bound = FALSE,
-                        folded = FALSE, normalized = FALSE, ...) {
+                        bound = FALSE,
+                        folded = NA, normalized = TRUE, ...) {
   check_flag(bound)
   check_unused(...)
 
@@ -123,8 +125,8 @@ rhat.mcmcrs <- function(x, by = "all", as_df = FALSE,
     x <- lapply(x, collapse_chains)
     x <- Reduce(bind_chains, x)
     return(rhat(x, by = by, as_df = as_df,
-                split = split, folded = folded, normalized = normalized))
+                folded = folded, normalized = normalized))
   }
   lapply(x, rhat, by = by, as_df = as_df,
-         split = split, folded = folded, normalized = normalized)
+         folded = folded, normalized = normalized)
 }